+++ oxygine/src/oxygine/actor/DebugActor.cpp
@@ -342,7 +342,7 @@ namespace oxygine
         s << "update=" << aligned(getStage()->_statUpdate, 2) << "ms ";
         s << "render=" << aligned(vstats.duration, 2) << "ms ";
         s << "textures=" << aligned(NativeTexture::created, 2) << " ";
-        s << "mats=" << aligned(mc().getTotalMaterials(), 2) << " ";
+        s << "mats=" << aligned((int)mc().getTotalMaterials(), 2) << " ";
 
 #ifdef __APPLE__
         size_t mem;
